requirements.txt
# Core dependencies for data generation
openai>=1.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Data processing and dataset management
datasets>=2.14.0
huggingface-hub>=0.17.0

# Training dependencies (for notebooks)
# Uncomment below if running training pipeline
# unsloth @ git+https://github.com/unslothai/unsloth.git
# xformers<0.0.27
# trl<0.9.0
# peft>=0.5.0
# accelerate>=0.21.0
# bitsandbytes>=0.41.0
# torch>=2.0.0
# transformers>=4.34.0
